+++ src/main/java/com/zpop/web/dao/MeetingDao.java
@@ -4,8 +4,8 @@ import com.zpop.web.dto.MeetingDetailDto;
 import com.zpop.web.dto.MeetingThumbnailPagination;
 import com.zpop.web.dto.ParticipantDto;
 import com.zpop.web.dto.UpdateMeetingViewDto;
-import com.zpop.web.dto.admin.AdminMeetingDetailsDto;
-import com.zpop.web.dto.admin.AdminMeetingDto;
+import com.zpop.web.dto.admin.meeting.AdminMeetingDetailsDto;
+import com.zpop.web.dto.admin.meeting.AdminMeetingDto;
 import com.zpop.web.entity.meeting.Meeting;
 import com.zpop.web.entity.meeting.MeetingThumbnailView;
 import org.apache.ibatis.annotations.Mapper;
