+++ src/main/java/com/zpop/web/dao/ReportedMeetingDao.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 
+import com.zpop.web.dto.admin.report.AdminReportedMeetingDto;
 import com.zpop.web.entity.ReportedMeeting;
-import com.zpop.web.dto.admin.AdminReportedMeetingDto;
 
 @Mapper
 public interface ReportedMeetingDao {
