--- src/main/java/com/zpop/web/dto/admin/AdminMemberDto.java
@@ -1,5 +1,0 @@
-package com.zpop.web.dto.admin;
-
-public class AdminMemberDto {
-
-}
