+++ src/main/java/com/zpop/web/dto/admin/comment/AdminCommentListResponse.java
@@ -1,4 +1,4 @@
-package com.zpop.web.dto.admin;
+package com.zpop.web.dto.admin.comment;
 
 import java.util.List;
