+++ src/main/java/com/zpop/web/dto/admin/member/AdminMemberDto.java
@@ -1,0 +1,5 @@
+package com.zpop.web.dto.admin.member;
+
+public class AdminMemberDto {
+
+}
