+++ src/main/java/com/zpop/web/service/admin/DefaultAdminReportService.java
@@ -15,9 +15,9 @@ import com.zpop.web.dao.MemberDao;
 import com.zpop.web.dao.ReportedCommentDao;
 import com.zpop.web.dao.ReportedMeetingDao;
 import com.zpop.web.dao.ReportedMemberDao;
-import com.zpop.web.dto.admin.AdminReportedCommentDto;
-import com.zpop.web.dto.admin.AdminReportedMeetingDto;
-import com.zpop.web.dto.admin.AdminReportedMemberDto;
+import com.zpop.web.dto.admin.report.AdminReportedCommentDto;
+import com.zpop.web.dto.admin.report.AdminReportedMeetingDto;
+import com.zpop.web.dto.admin.report.AdminReportedMemberDto;
 import com.zpop.web.entity.ReportedComment;
 import com.zpop.web.entity.ReportedMeeting;
 import com.zpop.web.entity.ReportedMember;
